fix(ci): maintainability — split oversized git_manager under line limit
Extract detached trusted-base worktree materialization and HEAD-object
probes into sibling modules so git_manager.py stays within the 1500-line
first-party maintainability gate without changing behavior.
